refactor: ♻️ replace showcase photo endpoint with simplified photos API.
Removes /api/photos/showcase.get.ts and fetch-masonry-media.ts. Replaces
them with /api/photos.get.ts and fetch-photos.ts, which fetches all
pages up to Pexels' 80-per-page max instead of using random page
selection with fallback filling.
